+++ src/main/java/ru/vitalib/otus/homework/books/dao/AuthorDao.java
@@ -1,18 +1,11 @@
 package ru.vitalib.otus.homework.books.dao;
 
-import java.util.List;
+import java.util.Optional;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
 import ru.vitalib.otus.homework.books.domain.Author;
 
-public interface AuthorDao {
-  Author save(Author author);
-
-  Author findById(long id);
-
-  void delete(Author author);
-
-  long count();
-
-  List<Author> findAll();
-
-  Author findByName(String authorName);
+@Repository
+public interface AuthorDao extends JpaRepository<Author, Long> {
+  Optional<Author> findByName(String authorName);
 }
